+++ src/components/Inputs/TextInput/TextInput.tsx
@@ -120,14 +120,29 @@ export const TextInput: FC<TextInputProps> = ({
         setClearButtonShown(false);
     };
 
-    const handleChange = useDebounce<React.ChangeEvent<HTMLInputElement>>(
+    const debouncedChange = useDebounce<React.ChangeEvent<HTMLInputElement>>(
         (
             _event?: React.ChangeEvent<HTMLTextAreaElement | HTMLInputElement>
         ) => {
-            onChange;
+            const { target } = _event;
+
+            onChange?.(_event);
+
+            if (target?.value.length === 0 && clearButtonShown) {
+                setClearButtonShown(false);
+            } else if (!clearButtonShown) {
+                setClearButtonShown(true);
+            }
         },
         waitInterval
     );
+
+    // We need to persist the syntheticevent object, as useDebounce uses a timeout function internally
+    // Reference: https://reactjs.org/docs/legacy-event-pooling.html
+    const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+        e.persist();
+        debouncedChange(e);
+    };
 
     return (
         <div className={textInputWrapperClassNames}>
@@ -142,25 +157,7 @@ export const TextInput: FC<TextInputProps> = ({
                 maxLength={maxlength}
                 minLength={minlength}
                 name={name}
-                onChange={
-                    !allowDisabledFocus
-                        ? (
-                              _event?: React.ChangeEvent<
-                                  HTMLTextAreaElement | HTMLInputElement
-                              >
-                          ) => {
-                              handleChange;
-                              if (
-                                  _event.target.value.length === 0 &&
-                                  clearButtonShown
-                              ) {
-                                  setClearButtonShown(false);
-                              } else if (!clearButtonShown) {
-                                  setClearButtonShown(true);
-                              }
-                          }
-                        : null
-                }
+                onChange={!allowDisabledFocus ? handleChange : null}
                 onBlur={!allowDisabledFocus ? onBlur : null}
                 onFocus={!allowDisabledFocus ? onFocus : null}
                 onKeyDown={!allowDisabledFocus ? onKeyDown : null}
@@ -205,7 +202,7 @@ export const TextInput: FC<TextInputProps> = ({
                     htmlType={iconButtonProps.htmlType}
                 />
             )}
-            {clearButtonShown && (
+            {clearButtonShown && !numbersOnly && htmlType !== 'number' && (
                 <DefaultButton
                     allowDisabledFocus={allowDisabledFocus}
                     ariaLabel={clearButtonAriaLabel}