+++ src/components/Modal/Modal.tsx
@@ -24,11 +24,13 @@ export const Modal: FC<ModalProps> = React.forwardRef(
             { [styles.small]: size === ModalSize.small },
             { [styles.medium]: size === ModalSize.medium },
             { [styles.large]: size === ModalSize.large },
+            { [styles.xLarge]: size === ModalSize.xLarge },
             { [styles.fullscreen]: size === ModalSize.fullscreen },
         ]);
 
         const headerClasses: string = mergeClasses([
             styles.header,
+            'header4',
             headerClassNames,
         ]);