+++ src/components/Pagination/index.ts
@@ -1,0 +1,3 @@
+export * from './Pager';
+export * from './Pagination';
+export * from './Pagination.types';