+++ src/hooks/useScrollLock.ts
@@ -1,23 +1,26 @@
-import React from 'react';
+import React, { useEffect, useRef } from 'react';
 
 /**
  * Hook to lock scroll on the given element
  * @param element
+ * @param shouldLock
  */
-export const useScrollLock = (element: HTMLElement) => {
-    let originalOverflow: string;
+export const useScrollLock = (element: HTMLElement, shouldLock: boolean) => {
+    let originalOverflow = useRef<string>('');
 
     const lockScroll = React.useCallback(() => {
-        originalOverflow = element.style.overflow;
+        originalOverflow.current = element.style.overflow;
         element.style.overflow = 'hidden';
     }, []);
 
     const unlockScroll = React.useCallback(() => {
-        element.style.overflow = originalOverflow;
+        element.style.overflow = originalOverflow.current;
     }, []);
 
-    return {
-        lockScroll,
-        unlockScroll,
-    };
+    useEffect(() => {
+        if (shouldLock) {
+            lockScroll();
+        }
+        return unlockScroll;
+    }, [element, shouldLock]);
 };