+++ src/shared/utilities.ts
@@ -110,5 +110,6 @@ export const stopPropagation = (e: React.MouseEvent<any>) =>
 /**
  * Get unique id
  */
-export const generateId = (): string =>
-    Math.random().toString(36).substring(2, 9);
+export const generateId = (prefix?: string) => {
+    return `${prefix ?? ''}${Math.random().toString(36).substring(2, 9)}`;
+};